+++ src/service/markdown/ext/markdown-it-katex.js
@@ -12,44 +12,10 @@ for rendering output.
 
 var katex = require('katex');
 
-// Test if potential opening or closing delimieter
-// Assumes that there is a "$" at state.src[pos]
-function isValidDelim(state, pos) {
-    var prevChar, nextChar,
-        max = state.posMax,
-        can_open = true,
-        can_close = true;
-
-    prevChar = pos > 0 ? state.src.charCodeAt(pos - 1) : -1;
-    nextChar = pos + 1 <= max ? state.src.charCodeAt(pos + 1) : -1;
-
-    // Check non-whitespace conditions for opening and closing, and
-    // check that closing delimeter isn't followed by a number
-    if (prevChar === 0x20/* " " */ || prevChar === 0x09/* \t */ ||
-            (nextChar >= 0x30/* "0" */ && nextChar <= 0x39/* "9" */)) {
-        can_close = false;
-    }
-    if (nextChar === 0x20/* " " */ || nextChar === 0x09/* \t */) {
-        can_open = false;
-    }
-
-    return {
-        can_open: can_open,
-        can_close: can_close
-    };
-}
-
 function math_inline(state, silent) {
     var start, match, token, res, pos, esc_count;
 
     if (state.src[state.pos] !== "$") { return false; }
-
-    res = isValidDelim(state, state.pos);
-    if (!res.can_open) {
-        if (!silent) { state.pending += "$"; }
-        state.pos += 1;
-        return true;
-    }
 
     // First check for and bypass all properly escaped delimieters
     // This loop will assume that the first leading backtick can not
@@ -79,14 +45,6 @@ function math_inline(state, silent) {
     if (match - start === 0) {
         if (!silent) { state.pending += "$$"; }
         state.pos = start + 1;
-        return true;
-    }
-
-    // Check for valid closing delimiter
-    res = isValidDelim(state, match);
-    if (!res.can_close) {
-        if (!silent) { state.pending += "$"; }
-        state.pos = start;
         return true;
     }