+++ src/service/markdown/ext/markdown-it-katex.js
@@ -23,14 +23,14 @@ function math_inline(state, silent) {
     // we have found an opening delimieter already.
     start = state.pos + 1;
     match = start;
-    while ( (match = state.src.indexOf("$", match)) !== -1) {
+    while ((match = state.src.indexOf("$", match)) !== -1) {
         // Found potential $, look for escapes, pos will point to
         // first non escape when complete
         pos = match - 1;
         while (state.src[pos] === "\\") { pos -= 1; }
 
         // Even number of escapes, potential closing delimiter found
-        if ( ((match - pos) % 2) == 1 ) { break; }
+        if (((match - pos) % 2) == 1) { break; }
         match += 1;
     }
 
@@ -49,8 +49,8 @@ function math_inline(state, silent) {
     }
 
     if (!silent) {
-        token         = state.push('math_inline', 'math', 0);
-        token.markup  = "$";
+        token = state.push('math_inline', 'math', 0);
+        token.markup = "$";
         token.content = state.src.slice(start, match);
     }
 
@@ -58,41 +58,41 @@ function math_inline(state, silent) {
     return true;
 }
 
-function math_block(state, start, end, silent){
+function math_block(state, start, end, silent) {
     var firstLine, lastLine, next, lastPos, found = false, token,
         pos = state.bMarks[start] + state.tShift[start],
         max = state.eMarks[start]
 
-    if(pos + 2 > max){ return false; }
-    if(state.src.slice(pos,pos+2)!=='$$'){ return false; }
+    if (pos + 2 > max) { return false; }
+    if (state.src.slice(pos, pos + 2) !== '$$') { return false; }
 
     pos += 2;
-    firstLine = state.src.slice(pos,max);
+    firstLine = state.src.slice(pos, max);
 
-    if(silent){ return true; }
-    if(firstLine.trim().slice(-2)==='$$'){
+    if (silent) { return true; }
+    if (firstLine.trim().slice(-2) === '$$') {
         // Single line expression
         firstLine = firstLine.trim().slice(0, -2);
         found = true;
     }
 
-    for(next = start; !found; ){
+    for (next = start; !found;) {
 
         next++;
 
-        if(next >= end){ break; }
+        if (next >= end) { break; }
 
-        pos = state.bMarks[next]+state.tShift[next];
+        pos = state.bMarks[next] + state.tShift[next];
         max = state.eMarks[next];
 
-        if(pos < max && state.tShift[next] < state.blkIndent){
+        if (pos < max && state.tShift[next] < state.blkIndent) {
             // non-empty line with negative indent should stop the list:
             break;
         }
 
-        if(state.src.slice(pos,max).trim().slice(-2)==='$$'){
-            lastPos = state.src.slice(0,max).lastIndexOf('$$');
-            lastLine = state.src.slice(pos,lastPos);
+        if (state.src.slice(pos, max).trim().slice(-2) === '$$') {
+            lastPos = state.src.slice(0, max).lastIndexOf('$$');
+            lastLine = state.src.slice(pos, lastPos);
             found = true;
         }
 
@@ -103,9 +103,9 @@ function math_block(state, start, end, silent){
     token = state.push('math_block', 'math', 0);
     token.block = true;
     token.content = (firstLine && firstLine.trim() ? firstLine + '\n' : '')
-    + state.getLines(start + 1, next, state.tShift[start], true)
-    + (lastLine && lastLine.trim() ? lastLine : '');
-    token.map = [ start, state.line ];
+        + state.getLines(start + 1, next, state.tShift[start], true)
+        + (lastLine && lastLine.trim() ? lastLine : '');
+    token.map = [start, state.line];
     token.markup = '$$';
     return true;
 }
@@ -113,42 +113,42 @@ function math_block(state, start, end, silent){
 module.exports = function math_plugin(md, options) {
     // Default options
 
-    options = options || {};
+    options = { throwOnError: false, strict: false };
 
     // set KaTeX as the renderer for markdown-it-simplemath
-    var katexInline = function(latex){
+    var katexInline = function (latex) {
         options.displayMode = false;
-        try{
+        try {
             return katex.renderToString(latex, options);
         }
-        catch(error){
-            if(options.throwOnError){ console.log(error); }
+        catch (error) {
+            if (options.throwOnError) { console.log(error); }
             return latex;
         }
     };
 
-    var inlineRenderer = function(tokens, idx){
+    var inlineRenderer = function (tokens, idx) {
         return katexInline(tokens[idx].content);
     };
 
-    var katexBlock = function(latex){
+    var katexBlock = function (latex) {
         options.displayMode = true;
-        try{
+        try {
             return "<p>" + katex.renderToString(latex, options) + "</p>";
         }
-        catch(error){
-            if(options.throwOnError){ console.log(error); }
+        catch (error) {
+            if (options.throwOnError) { console.log(error); }
             return latex;
         }
     }
 
-    var blockRenderer = function(tokens, idx){
-        return  katexBlock(tokens[idx].content) + '\n';
+    var blockRenderer = function (tokens, idx) {
+        return katexBlock(tokens[idx].content) + '\n';
     }
 
     md.inline.ruler.after('escape', 'math_inline', math_inline);
     md.block.ruler.after('blockquote', 'math_block', math_block, {
-        alt: [ 'paragraph', 'reference', 'blockquote', 'list' ]
+        alt: ['paragraph', 'reference', 'blockquote', 'list']
     });
     md.renderer.rules.math_inline = inlineRenderer;
     md.renderer.rules.math_block = blockRenderer;