+++ changelog.md
@@ -5,7 +5,8 @@
 - Fix preview html unable to load images.
 - Markdown:
   - Fix hr can not display on dark theme.
-  - Fixed export pdf not rendering math formulas that start or end with spaces.
+  - Edit math formulas using different background colors.
+  - Fix export pdf not rendering math formulas that start or end with spaces.
 
 # 2.8.0 2022-10-24