+++ changelog.md
@@ -1,6 +1,6 @@
 # Change log
 
-# 2.8.1 2022-10-28
+# 2.8.1 2022-10-29
 
 - Fix preview html unable to load images.
 - Markdown: